perf(point): store TimeOffset as ushort to halve per-point storage
Every drawn point is persisted in the append-only event log, so Point.
TimeOffset is one of the highest-cardinality fields in the database. It was
a long (BSON Int64, 8 bytes) even though it only needs to express
milliseconds since stroke start.

Strokes never need to exceed a minute and sub-millisecond precision is not
required, so a ushort (0..65535 ms, ~65.5 s) comfortably covers the domain.
The MongoDB driver serializes it as BSON Int32, halving the stored size per
point.

Update the model, its request/response DTOs, and clamp the client-computed
offset to 65535 so an unusually long stroke can't emit a value the narrower
type would reject. Adjust the plan annotations to match.

+++ Dtos/PointInput.cs
@@ -3,4 +3,4 @@ namespace Canvas.Dtos;
 /// <summary>
 /// Represents a point sent from the client when drawing a stroke.
 /// </summary>
-public sealed record PointInput(double X, double Y, double? Pressure, long TimeOffset);
+public sealed record PointInput(double X, double Y, double? Pressure, ushort TimeOffset);
